PC to Laptop

+++ src/main/java/tlcm/website/thebrewery/entities/material/Material.java
@@ -1,6 +1,7 @@
 package tlcm.website.thebrewery.entities.material;
 
 import javax.persistence.*;
+import java.math.BigInteger;
 
 @Entity
 @Table(name = "material")
@@ -8,7 +9,7 @@ public class Material {
     @Id
     @GeneratedValue
     @Column(name = "id")
-    private Integer id;
+    private BigInteger id;
 
     @Enumerated(value = EnumType.ORDINAL)
     @Column(name = "material_type")
