+++ src/main/java/tlcm/website/thebrewery/repository/MaterialRepository.java
@@ -6,12 +6,12 @@ import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 import tlcm.website.thebrewery.entities.material.Material;
 
+import java.math.BigInteger;
 import java.util.ArrayList;
-import java.util.List;
 
 @Repository
-public interface MaterialRepository extends JpaRepository<Material, Long> {
+public interface MaterialRepository extends JpaRepository<Material, BigInteger> {
 
     @Query(value = "SELECT m FROM Material m WHERE m.alcoholId = :alcoholId")
-    public ArrayList<Material> findAllByAlcoholId(@Param(value = "alcoholId") int alcoholId);
+    public ArrayList<Material> findAllByAlcoholId(@Param(value = "alcoholId") BigInteger alcoholId);
 }
